# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.ZmanVersion:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      formula_dsl:
        type: string
      id:
        type: string
      publisher_zman_id:
        type: string
      version_number:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zmanim.com
    name: Zmanim Lab Support
  description: API for Jewish prayer times calculation and publisher management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zmanim.com/terms
  title: Zmanim Lab API
  version: "1.0"
paths:
...
  /api/v1/publisher/zmanim/{zmanKey}/history:
    get:
      parameters:
      - description: Zman key
        in: path
        name: zmanKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.ZmanVersion'
            type: array
      summary: Get version history for a zman
      tags:
      - Publisher Zmanim
  /api/v1/publisher/zmanim/{zmanKey}/history/{version}:
    get:
      parameters:
      - description: Zman key
        in: path
        name: zmanKey
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ZmanVersion'
      summary: Get specific version of a zman
      tags:
      - Publisher Zmanim
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token from Clerk authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"